requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.11
fastapi-utils==0.8.0
h11==0.14.0
idna==3.10
mypy-extensions==1.0.0
psutil==5.9.8
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.34.0
